BF signed/unsigned mismatch

+++ iModelCore/ecobjects/src/ECSchemaConverter.cpp
@@ -871,7 +871,7 @@ IECInstanceR sourceCustomAttribute
 
     for (size_t i = 0; i < sourceCustomAttributeClass->GetPropertyCount(); i++)
         {
-        ECPropertyP propP = sourceCustomAttributeClass->GetPropertyByIndex(i);
+        ECPropertyP propP = sourceCustomAttributeClass->GetPropertyByIndex((uint32_t)i);
         ConvertPropertyValue(propP->GetName(), sourceCustomAttribute, *targetAttributeInstance);
         }
